Move demangle to detail
