+++ src/eckit/codec/detail/demangle.h
