+++ src/eckit/codec/types/StdVectorAdaptor.h
@@ -16,10 +16,10 @@
 
 #include "eckit/codec/Data.h"
 #include "eckit/codec/Metadata.h"
+#include "eckit/codec/detail/demangle.h"
 #include "eckit/codec/detail/TypeTraits.h"
 #include "eckit/codec/types/ArrayMetadata.h"
 #include "eckit/codec/types/ArrayReference.h"
-#include "eckit/codec/types/demangle.h"
 #include "eckit/exception/Exceptions.h"
 
 namespace std {
